# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Takomo
  version: "1.0.0-draft1"
  description: |
    Central task store for agent fleets. One authority, bearer-token auth,
    hierarchical tickets, enforced per-project workflows, atomic claim/lease,
    append-only event log.

    Design rules this contract encodes:
    - Most writes are commutative (comments, field sets, metadata merge); CAS
      (`If-Match` on `version`) is only needed for whole-body replacement.
    - Every rejection teaches: error bodies state what is allowed and how to
      proceed (`allowed_transitions`, `remedy`).
    - Claiming is one atomic operation (`POST /ready/claim`), never read-then-claim.
    - The event log (`global_seq`) is the only push/poll surface; all delivery
      mechanisms (SSE, hooks, channels) layer on the `since` cursor.
    - `/v1` is frozen: additive evolution only; experiments live in `metadata`.
servers:
  - url: /v1
security:
  - bearer: []
...
paths:
...
  /tokens:
    get:
      tags: [identity]
      summary: List token metadata (admin scope)
      description: |
        Returns metadata only — never the plaintext token or its at-rest hash.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Token" } }
        "403":
          description: Token lacks the admin scope
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      tags: [identity]
      summary: Mint a token over HTTP (admin scope)
      description: |
        Reuses the same store logic as the `takomo token create` CLI: the
        plaintext is returned exactly ONCE in the response and only its SHA-256
        is stored. This is a deliberate posture shift from auth.md's original
        "minting requires shell access" stance — see auth.md for the rationale.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenCreate" }
      responses:
        "201":
          description: Minted; plaintext shown once
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenCreated" }
        "403":
          description: Token lacks the admin scope
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Invalid body (missing scopes, bad projects, etc.)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /tokens/{id}:
    delete:
      tags: [identity]
      summary: Revoke a token by its public id (admin scope)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, description: token id, e.g. "tok_a8f2k1x9" }
      responses:
        "204": { description: Revoked }
        "403":
          description: Token lacks the admin scope
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/NotFound" }
  /projects:
    get:
      tags: [admin]
      summary: List projects
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Project" } }
    post:
      tags: [admin]
      summary: Create a project (admin scope)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /projects/{project}:
    delete:
      tags: [admin]
      summary: Delete a project and everything under it (admin scope)
      description: |
        Cascade-deletes the project together with ALL of its tickets, comments,
        dependency edges, and events, in a single transaction. Dependency edges
        that point INTO this project from another project's tickets are cleared
        too. A store-level `project_deleted` audit event (with a null `project`,
        so per-project event queries stay empty) records the deletion.

        Refused with `409 project.active_claims` when any ticket still holds an
        active (unexpired) claim — deleting under a live lease would yank work
        out from under a running worker. Pass `?force=true` to delete anyway,
        abandoning those leases.

        Tokens scoped to this project are deliberately left as-is: once the
        project is gone they simply stop resolving against it. Revoke them
        separately with `DELETE /v1/tokens/{id}` if desired.
      parameters:
        - { $ref: "#/components/parameters/project" }
        - name: force
          in: query
          required: false
          schema: { type: boolean, default: false }
          description: delete even when tickets hold active claims (abandons those leases)
      responses:
        "204": { description: Deleted }
        "403":
          description: Token lacks the admin scope
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: A ticket holds an active claim and force was not set
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /projects/{project}/workflow:
    get:
      tags: [workflow]
      summary: Get the project's workflow definition
      parameters: [ { $ref: "#/components/parameters/project" } ]
      responses:
        "200":
          description: The active workflow (see workflow-format.md)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workflow" }
    put:
      tags: [workflow]
      summary: Replace the project's workflow (admin scope; must remain valid for existing tickets)
      parameters: [ { $ref: "#/components/parameters/project" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Workflow" }
      responses:
        "200": { description: Replaced }
        "422":
          description: Invalid workflow (unreachable states, tickets stranded)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /projects/{project}/roadmap:
    get:
      tags: [workflow]
      summary: Epic progress rollup (read scope)
      description: |
        For each `epic` in the project, returns the epic plus a rollup over its
        FULL descendant subtree (walked recursively via `parent`): ticket counts
        by state and by category, the subtree `total`, a `done` count (states in
        the `done` category), and a whole-number completion `percent`. The epic
        itself is the container and is not counted in its own rollup. Counts
        include archived descendants so historical completion is not understated.
      parameters: [ { $ref: "#/components/parameters/project" } ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  project: { type: string }
                  generated_at: { type: string, format: date-time }
                  epics:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        title: { type: string }
                        state: { type: string }
                        state_category: { type: string }
                        priority: { type: string }
                        total: { type: integer, description: descendants in the subtree }
                        done: { type: integer, description: descendants in a done-category state }
                        percent: { type: integer, description: done/total as a whole percent (0 when empty) }
                        by_state: { type: object, additionalProperties: { type: integer } }
                        by_category: { type: object, additionalProperties: { type: integer } }
        "404": { $ref: "#/components/responses/NotFound" }
  /tickets:
    post:
      tags: [tickets]
      summary: Create a ticket
      description: |
        `Idempotency-Key` is required for agent clients: retries with the same
        key return the original ticket (200) instead of creating a twin (201).
        The response includes `similar` — existing open tickets with matching
        title keywords — so a client can notice a semantic duplicate.
      parameters:
        - { name: Idempotency-Key, in: header, required: false, schema: { type: string, maxLength: 128 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TicketCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TicketCreated" }
        "200":
          description: Idempotent replay of a prior create
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TicketCreated" }
        "422":
          description: Validation failure (unknown project, bad parent, cycle)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [tickets]
      summary: List/search tickets
      parameters:
        - { name: project, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string }, description: state id or category }
        - { name: type, in: query, schema: { type: string } }
        - { name: label, in: query, schema: { type: string }, description: repeatable, AND semantics }
        - { name: parent, in: query, schema: { type: string }, description: direct children of this id }
        - { name: q, in: query, schema: { type: string }, description: full-text over title+body }
        - { name: claimed_by, in: query, schema: { type: string } }
        - { name: fields, in: query, schema: { type: string }, description: comma list for sparse responses }
        - { name: include_archived, in: query, schema: { type: boolean, default: false }, description: include archived tickets alongside active ones }
        - { name: archived, in: query, schema: { type: string, enum: [only, all, active] }, description: "only = archived only; all = both; active = default (non-archived)" }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Ticket" } }
                  next_cursor: { type: [string, "null"] }
  /tickets/{id}:
    get:
      tags: [tickets]
      summary: Get a ticket
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: include, in: query, schema: { type: string }, description: "comma list: comments,children,deps,events" }
      responses:
        "200":
          description: OK
          headers:
            ETag: { schema: { type: string }, description: current version, e.g. "7" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [tickets]
      summary: Update fields (commutative; CAS only for body)
      description: |
        Field-level sets (`title`, `priority`, `labels`) and `metadata_merge`
        (RFC 7386 merge-patch semantics, namespaced keys) are commutative and
        never require `If-Match`. `links` MERGES per key: each key in the patch
        is set (string value) or deleted (null value) while untouched keys
        survive, so concurrent writers attaching different link keys never
        clobber each other. Replacing `body` requires `If-Match: "<version>"`;
        on mismatch the 409 carries the current version and body hash. State is
        NOT patchable — use /transition.
        While a ticket is claimed, non-holders may only patch `metadata_merge`
        under their own namespace and add comments; other fields 409.
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: If-Match, in: header, required: false, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TicketPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "409": { $ref: "#/components/responses/Conflict" }
        "404": { $ref: "#/components/responses/NotFound" }
  /tickets/{id}/transition:
    post:
      tags: [workflow]
      summary: Move a ticket to another state (the only way state changes)
      description: |
        Validated against the project workflow: transition must exist from the
        current state, all `requires` must hold (active claim held by caller,
        token scope, guards like no-open-children). Illegal transitions return
        409 with `allowed_transitions` and a `remedy` — errors are written for
        LLM readers.

        Checks run in the order legality → scope → claim/fence, and the headline
        `code`/`message` names the first real blocker in that order: an illegal
        target is `transition.illegal`, a missing scope is `transition.scope`
        (403), and only then are claim ownership and fence echoing checked. So a
        lease holder attempting a human gate they lack the scope for sees the
        403, not a fencing error.

        Human approval is authoritative over a lease: a transition whose
        `requires` includes `scope:human`, performed by a caller holding the
        `human` scope, succeeds even while another actor holds the claim, and
        auto-releases that claim as a side effect (emitting a `released` event
        alongside `transitioned`). This override is scoped to human-required
        transitions; ordinary `claim`-required transitions keep the holder lock.
      parameters: [ { $ref: "#/components/parameters/id" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to: { type: string, description: target state id }
                reason: { type: string, maxLength: 2000 }
                fence: { type: integer, description: required while ticket is claimed }
      responses:
        "200":
          description: Transitioned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "409":
          description: Illegal transition, missing claim, stale fence, or guard failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransitionError" }
        "403":
          description: Transition requires a scope this token lacks (e.g. human)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransitionError" }
  /tickets/{id}/comments:
    post:
      tags: [tickets]
      summary: Append a comment (always allowed, commutative)
      parameters: [ { $ref: "#/components/parameters/id" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, maxLength: 65536 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" }
  /tickets/{id}/deps:
    get:
      tags: [tickets]
      summary: The dependency graph around a ticket (reverse + transitive, cycle-safe)
      description: |
        Returns the dependency graph reachable from the ticket. `direction`
        selects which edges to follow: `blocked_by` (default) walks to the
        tickets that block this one; `blocks` walks the inverse (the tickets
        this one blocks); `both` walks both. With `transitive=true` the walk
        chases the whole chain to its ends; otherwise only the ticket's direct
        neighbours are returned. The traversal is cycle-safe (a visited set
        bounds it even when A blocks B blocks A). Edges are canonical
        `{ticket, blocked_by}` pairs (ticket is blocked by blocked_by)
        regardless of traversal direction, so `both` never duplicates an edge.
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: direction, in: query, schema: { type: string, enum: [blocked_by, blocks, both], default: blocked_by } }
        - { name: transitive, in: query, schema: { type: boolean, default: false } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  root: { type: string }
                  direction: { type: string, enum: [blocked_by, blocks, both] }
                  transitive: { type: boolean }
                  nodes:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        title: { type: string }
                        state: { type: string }
                        state_category: { type: string }
                        type: { type: string }
                  edges:
                    type: array
                    items:
                      type: object
                      properties:
                        ticket: { type: string, description: the blocked ticket }
                        blocked_by: { type: string, description: the ticket that blocks it }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [tickets]
      summary: Add a blocked_by edge (cycle-checked)
      description: |
        Dependency edges drive readiness and the no_open_blockers guard, so
        they follow the claimed-ticket write rules: while the ticket is
        claimed, only the holder may change deps and must echo the lease
        `fence` (per the Lease schema's fence-on-every-mutating-call rule).
      parameters: [ { $ref: "#/components/parameters/id" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [blocked_by]
              properties:
                blocked_by: { type: string, description: ticket id that blocks this one }
                fence: { type: integer, description: required while ticket is claimed }
      responses:
        "201": { description: Added }
        "409": { $ref: "#/components/responses/Conflict" }
        "422":
          description: Would create a cycle, or unknown id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [tickets]
      summary: Remove a blocked_by edge
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: blocked_by, in: query, required: true, schema: { type: string } }
        - { name: fence, in: query, required: false, schema: { type: integer }, description: required while ticket is claimed }
      responses:
        "204": { description: Removed }
        "409": { $ref: "#/components/responses/Conflict" }
  /tickets/{id}/archive:
    post:
      tags: [tickets]
      summary: Archive a ticket (hide from default views; write scope)
      description: |
        Sets `archived_at` so the ticket drops out of the default list, ready
        queue, board, and metrics. The ready queue and metrics always exclude
        archived tickets; `GET /v1/tickets` re-includes them with
        `include_archived=true` or `archived=only`. Any state may be archived —
        terminal done/cancelled is the typical case. Idempotent.
      parameters: [ { $ref: "#/components/parameters/id" } ]
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "404": { $ref: "#/components/responses/NotFound" }
  /tickets/{id}/unarchive:
    post:
      tags: [tickets]
      summary: Unarchive a ticket (return to default views; write scope)
      parameters: [ { $ref: "#/components/parameters/id" } ]
      responses:
        "200":
          description: Unarchived
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "404": { $ref: "#/components/responses/NotFound" }
  /tickets/{id}/claim:
    post:
      tags: [claims]
      summary: Claim this specific ticket (atomic)
      description: |
        Succeeds only if the ticket is in a claimable state, unclaimed (or the
        caller already holds it — idempotent renewal), and unblocked. Returns
        the lease with the fencing token all subsequent mutating calls must echo.
      parameters: [ { $ref: "#/components/parameters/id" } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ttl_seconds: { type: integer, default: 900, maximum: 3600 }
      responses:
        "200":
          description: Claimed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lease" }
        "409":
          description: Already claimed by another holder, blocked, or state not claimable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /tickets/{id}/heartbeat:
    post:
      tags: [claims]
      summary: Renew the lease (harness/orchestrator responsibility, not the model's)
      parameters: [ { $ref: "#/components/parameters/id" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fence]
              properties:
                fence: { type: integer }
                ttl_seconds: { type: integer, default: 900, maximum: 3600 }
      responses:
        "200":
          description: Renewed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lease" }
        "409":
          description: |
            Fence mismatch. `fence.stale` (presented < current) means the lease
            expired and was superseded by a later claim — stop writing.
            `fence.invalid` (presented > current, a value never issued) is a
            client bug — re-read the fence from your claim response.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /tickets/{id}/release:
    post:
      tags: [claims]
      summary: Voluntarily release the claim
      parameters: [ { $ref: "#/components/parameters/id" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fence]
              properties:
                fence: { type: integer }
                reason: { type: string }
      responses:
        "204": { description: Released }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /events:
    get:
      tags: [events]
      summary: Read the event log from a cursor (the only source of truth for "what changed")
      parameters:
        - { name: since, in: query, required: true, schema: { type: integer }, description: last seen global_seq; 0 for from-start }
        - { name: project, in: query, schema: { type: string } }
        - { name: ticket, in: query, schema: { type: string } }
        - { name: kind, in: query, schema: { type: string }, description: comma list of event kinds }
        - { name: wait, in: query, schema: { type: integer, default: 0, maximum: 120 }, description: long-poll seconds }
        - { name: limit, in: query, schema: { type: integer, default: 200, maximum: 1000 } }
      responses:
        "200":
          description: Events after `since`, oldest first; empty array on wait timeout
          content:
            application/json:
              schema:
                type: object
                properties:
                  events: { type: array, items: { $ref: "#/components/schemas/Event" } }
                  cursor: { type: integer, description: pass as next `since` }
  /events/stream:
    get:
      tags: [events]
      summary: SSE stream of events (doorbell; cursor remains the source of truth)
      description: |
        `text/event-stream`. Each event's `id:` is its global_seq, so
        `Last-Event-ID` resumes seamlessly. Consumers must treat the stream as
        lossy and reconcile via GET /events on reconnect.
      parameters:
        - { name: since, in: query, schema: { type: integer } }
        - { name: project, in: query, schema: { type: string } }
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema: { type: string }
  /export:
    get:
      tags: [admin]
      summary: Export tickets (with comments and deps) as JSONL — read scope
      description: |
        Streams one JSON object per line (`application/x-ndjson`): each line is
        a full ticket (its `blocked_by` array carries the dependency edges) plus
        a `comments` array. `metadata` and `links` ride along verbatim, so the
        export is a faithful snapshot. Round-trips through the `takomo import
        --from takomo` CLI verb. Honors token project scoping; `project`
        narrows to one project.
      parameters:
        - { name: project, in: query, schema: { type: string }, description: narrow to one project (else all readable projects) }
      responses:
        "200":
          description: JSONL stream, one ticket object per line
          content:
            application/x-ndjson:
              schema: { type: string }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: See auth.md — `tk_`-prefixed tokens, scoped, hashed at rest.
    shareBearer:
      type: http
      scheme: bearer
      description: |
        A SHARE token (`tks_`-prefixed), minted by `POST /v1/shares` and hashed
        at rest like a normal token. It is a distinct auth path: it authorizes
        ONLY the read-only `/v1/shares/self*` endpoints, bounded to its scope,
        and is rejected on every normal endpoint. Expired/revoked shares return
        `410 Gone`. See auth.md.
...
  responses:
    NotFound:
      description: Unknown id
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: "stable machine code, e.g. conflict.version, claim.held, fence.stale (superseded lease), fence.invalid (fence never issued — a client bug)" }
        message: { type: string, description: written for LLM readers — states what is wrong AND what to do }
        remedy: { type: string, description: the exact next call to make, when one exists }
        current_version: { type: integer }
        details: { type: object }
...